Move global components to ui package
