+++ apps/dataset-browser/.eslintrc.js
@@ -1,4 +1,12 @@
 module.exports = {
   root: true,
   extends: ['custom'],
+  rules: {
+    'node/no-extraneous-import': [
+      'error',
+      {
+        allowModules: ['ui'],
+      },
+    ],
+  },
 };
