+++ apps/dataset-browser/src/app/[locale]/contact/page.tsx
@@ -1,4 +1,4 @@
-import LocalizedMarkdown from '@/components/localized-markdown';
+import {LocalizedMarkdown} from 'ui';
 
 export default function Contact() {
   return <LocalizedMarkdown name="contact" />;
