+++ apps/dataset-browser/src/app/[locale]/faq/page.tsx
@@ -1,4 +1,4 @@
-import LocalizedMarkdown from '@/components/localized-markdown';
+import {LocalizedMarkdown} from 'ui';
 
 export default function Faq() {
   return <LocalizedMarkdown name="faq" />;
