+++ apps/dataset-browser/src/app/[locale]/navigation.tsx
