+++ packages/ui/index.tsx
@@ -1,2 +1,5 @@
 import * as React from 'react';
-export * from './Button';
+export * from './badge';
+export * from './page';
+export * from './localized-markdown';
+
